+++ processor/src/test/fixtures/bad_input/com/example/action/ReportParametersWithNoParametersActionModel.java
@@ -1,0 +1,13 @@
+package com.example.action;
+
+import arez.annotations.Action;
+import arez.annotations.ArezComponent;
+
+@ArezComponent
+public abstract class ReportParametersWithNoParametersActionModel
+{
+  @Action( reportParameters = false )
+  public void doStuff()
+  {
+  }
+}